\#8: swot serializers
